% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{praiseme}
\alias{praiseme}
\title{train the machine learning data used in the analysis of RNAseq data
Load the Maxquant outpu data into R environment}
\usage{
praiseme(your_name)
}
\value{
returns a list of genes
}
\description{
train the machine learning data used in the analysis of RNAseq data
Load the Maxquant outpu data into R environment
}
