% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcall.R
\name{splt}
\alias{splt}
\title{This is the package that is used in the analysis of gene sets}
\usage{
splt(gsea)
}
\arguments{
\item{gsea}{}
}
\value{
gene sets indentified
}
\description{
This is the package that is used in the analysis of gene sets
}
