% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcall.R
\name{varcall}
\alias{varcall}
\title{This is the package that is used in the analysis of gene sets}
\usage{
varcall(n)
}
\arguments{
\item{n}{}
}
\value{
gene sets indentified
}
\description{
This is the package that is used in the analysis of gene sets
}
